Music Play/Crossfade Callback
Adds the music callback hooks for the api. You can return a table with 2 params, an int to replace the track or a false to cancel. Theres a 3rd extra param to tell you if the func being called is crossfade or play, it's not ideal, but it's better than needing to add 2 callbacks for esscentially the same thing, in my eyes.