Fix generation of trampolines for virtual functions
Because virtual functions call will always compile to a vtable call, the
previous version of the code emitter would result in a direct call to the
vtable method but with the default calling convention, even if it should use
a different one.

In this new version, virtual functions are emitted as non virtual functions, but
the trampoline will call the virtual function regardless. This solves only part of
the problem, as it now prevents non virtual calls.