Improvements to ASM patching
  * Added the ByteBuffer class that can be used to represent arbitrary
sequences of bytes, even if they include zeroes. This effectively acts
as a goto solution when generating dynamic assembly, where zeroes may
appear in the middle of strings, resulting in some operations not
reading the entire patch, such as strcpy.
  * Added ASMPatch::AddBytes(ByteBuffer const& bytes) to use a ByteBuffer
in an ASMPatch.
  * Added ASMPatch::PreserveRegisters and ASMPatch::RestoreRegisters
that can be used to save / restore registers. Saved registers can be
set to critical (by default) and the runtime will throw an exception if
such registers have not been restored within the context of the patch.
  * Added ASMPatch::Push and ASMPatch::Pop that can be used to arbitrarily
push and pop general purpose registers to and from the stack
  * Added ASMPatch::MoveFromMemory and ASMPatch::MoveToMemory that can be
used to perform `mov dst, [src + offset]` or `mov [dst + offset], src`.
Kind reminder that memory-to-memory movs are not allowed in x86.
  * Added ASMPatch::LoadEffectiveAddress that implements a limited form of
the `lea` instruction, mainly the `lea dst, [src+offset]` version.