add new INI for options / vanilla tweaks
This adds a Repentogon/options.ini file in the save folder. Currently
this allows for toggling three of our tweaks, which is the Key Master
deal chance, Hush panic state fix, and better Void generation.
Eventually I want to hook this up to an options menu but that's future
me's problem.