player stat refactor
We no longer turn modded characters that declare custom stats into Eden.
This ended up causing many edge cases that had to manually be fixed
later. Instead of doing that, this functionality has been replaced with
six assembly patches that directly inject stats into the same places
where Eden's stats are calculated (with the exception of speed, as where
Eden's stat is added here isn't obvious in the disassembly so it's added
directly.