imgui: horizontal scroll's final-ish form
ImGui has been instructed to never manipulate the scroll position of the
console's textbox input. We are handling that ourselves. This is to
allow the user full control over the scrolling of the console.

This is not what I could call ideal, but it's functional and much better
than how the console used to be. The console reacts naturally to how I
feel it should, with no more unintuitive jerks in scrolling position.

This may need further refinement down the road, but works for now.