[Parser] Fix placement of implicit output ArgData
For non x86-64 w/ output functions, the implicit output arg data of
the function was emitted as the last arg data, causing it to be
incorrectly pushed on the stack first instead of last during the
generation of hooks / supers.

This patch fixes this, so now implicit output arg data is the first
in the array of arg data.

Additionally, the commit fixes the placement of commas in the array,
as they used to appear at the very end of each line, rather than
appearing after the definition of the ArgData (which could result
in invalid code under certain circumstances).