Fix stack corruption in ItemPool::load_pools
Summary: This fixes a stack corruption that occured as a result of
ASM patching out a call to KAGE::LogMessage. The stack was no longer
prepared as a jump was performed over the call to LogMessage, but
the stack was still cleaned as if it had been prepared.

This changes the patch to a simple nop replacement of the call to
LogMessage, allowing the stack to remain balanced.