Autoconfigure ZHL and Repentogon version symbols
Summary: This adds CMake variables that are exported as compile definitions
to define the version of ZHL and Repentogon. This variables are set to "dev"
by default, and are set to the content of version.txt in the appveyor script.
This ensures the value of the symbols remains up-to-date whenever Repentogon
releases a new version.