Two stages DLL mod loading
Summary: This replaces the previous mod loading process, where all
DLLs mods were loaded immediately after libzhl is loaded, to a two
stages process.

First libzhl is loaded and manually hooks in the middle of the main
function of Isaac. This hook will proceed with the load of the DLL
mods. This allows Repentogon (in particular) to properly update the
title of the window as its initialization process goes on.