Fix Mods_Menu::Render font rendering patch
Summary: This fixes the ASM patch that voids the calls to
DrawStringScaled in Mods_Menu::Render. Previously, the patch merely
jumped over the call to DrawStringScaled, however this resulted in
a stack unbalanc (that was unnoticed until now, somehow). This new
patch replaces the calls to DrawStringScaled with a simple balancing
of the stack.