Perform the second stage DLL loading after KAGE::StartupEngine
Summary: This fixes the issue where function hooks would not be
applied because the second stage DLL loading was performed too
late.

In the previous version, the hook to call the second stage DLL
loader was executed after calling IsaacStartup. However, IsaacStartup
initializes the Lua API, as well as EntityFactory, and as such some
hooks were not applied. KAGE::StartupEngine is the earlist moment
when we can both hook all functions and have a defined window we
can update as operations are performed.