Add hook enabler system
This adds the ability for FunctionDefinition in ZHL to indicate whether
they allow hooks to be applied on them or not. This is setup through a
JSON file (hooks.json) that contains pairs consisting of a name and a
boolean, indicating whether the FunctionDefinition with the given allows
hooks or not.

The ZHL parser now allows function definitions in ZHL files to receive
the "nohook" qualifier, which prevents a function from being hooked. The
parser also generate a JSON file containing the hook state of all
functions, as defined in ZHL files.

In order of priority:
  * As the parser is ran every time, the output IsaacRepentance.cpp file
    configures the hook status of every definition based on what is
    present in ZHL files. If a function is not changed by the JSON, then
    its hook status is not updated.
  * When the game starts, ZHL loads the JSON file. For every function it
    finds inside the JSON, it updates the hook status of the associated
    FunctionDefinition (if any exists).