random: process hidden chars before locked ones
While testing the tainted menu, I noticed an edge case thanks to
Community Remix, which currently has a character marked as hidden as it
is unfinished, but still has an achievement ID tied to it. This would
cause the character menu to desync during random selection.

As we do not have the capability of unlocking hidden characters, there
is no point in processing unlocking for them and they can be immediately
skipped without incrementing the random offset.