Establish separated REPENTOGON persistentgamedata save files & the functionality to automatically keep them in sync with the vanilla saves.
SaveFile: Class used to read the raw contents of a save file into memory, modify/synchronize it, then write it back to disk. Contains the logic used to actually read/modify the saves during synchronization.
SaveSyncing: Contains the patches/hook/logic for orchestrating save synchronization (when to do it, what mode to use, when not to do it, etc).