Add (hopefully temporary) startup safety
Force the Isaac exe to be ran through the launcher if Repentogon is
requested. In addition, do not allow the game to run if Repentogon is
not found.

This prevents an issue where the game breaks savefiles when forward
compatibility is not assured (e.g. running a v14 savefile on a v12
game).